backend/requirements.txt
arch==7.2.0
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
Cython==3.0.11
Django==4.2
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.14.0
frozendict==2.4.6
html5lib==1.1
idna==3.10
joblib==1.4.2
lxml==5.3.0
multitasking==0.0.11
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pandas-datareader==0.10.0
patsy==1.0.1
peewee==3.17.9
platformdirs==4.3.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
soupsieve==2.6
sqlparse==0.5.1
statsmodels==0.14.4
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
webencodings==0.5.1
yfinance==0.2.52
openai==1.62.0
